Address review comments and improve error reporting for `llvm_refine_spec`.
